--- FE/src/hooks/useProject.tsx
@@ -1,1 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
