+++ FE/src/pages/ProjectCreatePage.tsx
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+const ProjectCreatePage = (): ReactNode => {
+  return <div>ProjectCreatePage</div>;
+};
+
+export default ProjectCreatePage;
