+++ FE/src/pages/ProjectPage.tsx
@@ -1,12 +1,7 @@
-interface projectElement extends projectCardElement {
-  projectId: number;
-}
-
-interface projectCardElement {
-  title: string;
-  description: string;
-  taskNum: number;
-}
+import { projectElement } from '../types/project';
+import ProjectCard from '../components/project/ProjectCard';
+import whiteCross from '../assets/images/cross.png';
+import { Link } from 'react-router-dom';
 
 const dummyData: Array<projectElement> = [
   {
@@ -72,40 +67,30 @@ const dummyData: Array<projectElement> = [
   },
 ];
 
-const ProjectCard = ({ title, description, taskNum }: projectCardElement) => {
-  return (
-    <div className="w-[11.25rem] h-[7.5rem] bg-true-white px-3 pt-3 pb-6 border rounded-[0.25rem] border-transparent-green flex flex-col justify-between">
-      <div className="flex flex-col gap-1">
-        <div className="font-pretendard text-ml text-starbucks-green font-bold">{title}</div>
-        <div className="font-pretendard text-s text-true-black font-medium">{description}</div>
-      </div>
-      <div className="flex gap-1">
-        <div className="font-pretendard text-s text-true-black font-medium">내 진행 업무</div>
-        <div className="font-pretendard text-s bg-starbucks-green w-6 h-4 flex justify-center content-center rounded-[0.25rem]">
-          <div className="text-true-white text-s font-medium">{taskNum}</div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
 const ProjectPage = () => {
   return (
     <>
-      <div className="flex gap-2 justify-center min-w-[76rem] mt-8 mb-6">
+      <header className="flex gap-2 justify-center min-w-[76rem] mt-8 mb-6">
         <div className="bg-house-green w-[66.5rem] h-14 font-bold font-pretendard text-true-white text-l ps-3 flex flex-col justify-center rounded-[0.25rem]">
           내 프로젝트
         </div>
         <div className="bg-house-green w-[8.6875rem] h-14 font-bold font-pretendard text-true-white text-ml flex flex-col justify-center rounded-[0.25rem] text-center">
           로그아웃
         </div>
-      </div>
+      </header>
       <div className="flex flex-wrap gap-x-5 gap-y-4 w-[76rem] min-w-[76rem] mx-auto">
         {dummyData.map(({ title, description, taskNum, projectId }) => (
           <li key={projectId} className="list-none">
             <ProjectCard title={title} description={description} taskNum={taskNum} />
           </li>
         ))}
+        <Link
+          to={'/project/create'}
+          className="w-[7.5rem] h-[7.5rem] flex flex-col justify-center content-center gap-2 bg-starbucks-green rounded-[0.25rem] cursor-pointer"
+        >
+          <div className="text-true-white font-pretendard font-bold text-r text-center">새 프로젝트 생성</div>
+          <img src={whiteCross} className="w-4 h-4 mx-auto" />
+        </Link>
       </div>
     </>
   );
