+++ FE/src/types/project.ts
@@ -1,0 +1,9 @@
+export interface projectCardElement {
+  title: string;
+  description: string;
+  taskNum: number;
+}
+
+export interface projectElement extends projectCardElement {
+  projectId: number;
+}
