+++ FE/tailwind.config.js
@@ -22,11 +22,11 @@ export default {
         pretendard: ['pretendard'],
       },
       fontSize: {
-        s: '0.75rem',
-        r: '0.875rem',
-        m: '1rem',
-        ml: '1.125rem',
-        l: '1.5rem',
+        s: '12px',
+        r: '14px',
+        m: '16px',
+        ml: '18px',
+        l: '24px',
       },
     },
   },
